+++ tests/infrastructures/symfony/ContainerLaminasTest.php
@@ -1,0 +1,73 @@
+<?php
+
+/**
+ * East Website.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license and the version 3 of the GPL3
+ * license that are bundled with this package in the folder licences
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richarddeloge@gmail.com so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) 2009-2020 Richard Déloge (richarddeloge@gmail.com)
+ *
+ * @link        http://teknoo.software/east/website Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+
+namespace Teknoo\Tests\East\WebsiteBundle;
+
+use DI\Container;
+use DI\ContainerBuilder;
+use Laminas\Diactoros\StreamFactory;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class DefinitionProviderTest.
+ *
+ * @copyright   Copyright (c) 2009-2020 Richard Déloge (richarddeloge@gmail.com)
+ *
+ * @link        http://teknoo.software/east Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+class ContainerLaminasTest extends TestCase
+{
+    /**
+     * @return Container
+     */
+    protected function buildContainer() : Container
+    {
+        $containerDefinition = new ContainerBuilder();
+        $containerDefinition->addDefinitions(__DIR__.'/../../../infrastructures/symfony/Resources/config/laminas_di.php');
+        $containerDefinition->useAutowiring(false);
+        
+        return $containerDefinition->build();
+    }
+
+    public function testStreamFactory()
+    {
+        $container = $this->buildContainer();
+
+        self::assertInstanceOf(
+            StreamFactory::class,
+            $container->get(StreamFactory::class)
+        );
+    }
+
+    public function testMediaEndPointStreamFactory()
+    {
+        $container = $this->buildContainer();
+
+        self::assertInstanceOf(
+            StreamFactory::class,
+            $container->get('teknoo.east.website.endpoint.media.stream.factory')
+        );
+    }
+}