+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Teknoo Software - Website - Change Log
 
+##[3.1.1] - 2020-09-10
+###Stable Release
+###Update
+- Add Di definition dedicated to Laminas for Symfony bundle to work out of the box with the metapackage 
+  `teknoo/east-symfony-symfony`.
+
 ##[3.1.0] - 2020-09-10
 ###Stable Release
 ###Update